refactor binary puzzle solution generation Previous solution took about 24 seconds to generate every six by six solution. This one takes about 36 seconds to do the same, but it offers more flexibility in optimising.
The idea is to only visit squares built from valid rows. This greatly reduces
the amount of squares visited, at the cost of memory (we have to keep every possible
valid row in memory). The problem is that my implementation isn't very efficient.
Especially the list comprehensions seem to be taking a lot of time, so conversion
to a one-dimensional list so that we don't need to generate separate column lists
seems proper here. After that it might be worth considering
writing part of the code in C.

This implementation does not work well for larger puzzle sizes because of the memory footprint.
since this is binary, we might try using actual bits instead of multiple bytes per number.

We could also break the rows up into multiple blocks to be combined one at a time.